msm: fb: Acknowledge a current display state
Use LCD notifier to permanently store and share the current state
of a MDP panel. As there is no way to ensure the panel is off right
now via a simple interface, implement it within notifier call chain.
In case LCD notify is not enabled, lcd_panel_suspended() call will
return an appropriate negative, so this case must be handled by
clients.

Change-Id: I586fc40eed7f7c4ccc5a73a889778bbe0ad8afc8
Signed-off-by: Alex Saiko <solcmdr@gmail.com>